+++ libraries/base/changelog.md
@@ -15,6 +15,9 @@
   * Add `firstA` and `secondA` to `Data.Bitraversable`. ([CLC proposal #172](https://github.com/haskell/core-libraries-committee/issues/172))
   * Deprecate `GHC.TypeNats.Internal`, `GHC.TypeLits.Internal`, `GHC.ExecutionStack.Internal` ([CLC proposal #217](https://github.com/haskell/core-libraries-committee/issues/217))
   * Define `Eq1`, `Ord1`, `Show1` and `Read1` instances for basic `Generic` representation types. ([CLC proposal #273](https://github.com/haskell/core-libraries-committee/issues/273))
+  * Add exception type metadata to default exception handler output.
+    ([CLC proposal #231](https://github.com/haskell/core-libraries-committee/issues/231)
+    and [CLC proposal #261](https://github.com/haskell/core-libraries-committee/issues/261))
 
 ## 4.20.0.0 May 2024
   * Shipped with GHC 9.10.1